Complete step 4
This is distorted, of course. I need to figure out the most idiomatic
way to scale the scene so that the square is actually rendered as a
square, regardless of the proportions of the canvas. The hacky way to do
it would be to modify the raw vertex data, but I would prefer to learn
how to do it without doing that. This might be where the vertex shader
comes in, but I have some digging to do. The simplest solution for now
might be to just use CSS to force the canvas to be a square.