Pull draw code into a function
I might regret this. I was starting to go down the path of trying to
redraw when the aspect ratio changed with some kind of uniform. I’ve
given up on that and am going to move onto the next step in the
tutorial, but I’m keeping the `draw()` function as-is under the
assumption that eventually I’ll need a render loop of some kind anyway.